pyproject.toml
[project]
name = "eom"
version = "0.1.0"
description = "EdgeOMatic control library"
requires-python = ">=3.10"
dependencies = [
    "litestar",
    "granian",
    "msgspec",
    "websocket-client",
    "uvicorn",
]

[project.scripts]
eom-server = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "B", "I"] 
